Extend display_result func by only_person flag
The display_result function was extended by a further input of the type
bolean. This was implemened to fix the bug, where the label on the
bounding box was always displayed incorrectly as aeroplane when working
with the subset of the date set that only contains images of the class
"person".