Add GOOGLE_DRIVE flag
A new bool flag was added at the beginning of the notebook. If set
correctly according to whether the user is running the notebook localy
or on googe colab, the environment and paths are set automatically.
This enables the user to just set the flag and then run all cells
without the need to pay any further attention.