Add Kfold
The notbook was modifed to use Kfold to split the test data into a test
dataset and a validation set. The traditional method of splitting the
traning data wasn't applied, as the training dataset would be smaller
than the test dataset. Moreover, further flags were added allow the use
of the already existing way of using custiom splits. For clarity some
existing functions had to be renamed.