Export model to onnx and execute measurements
The tinyyolov2_fused model was converted to the ONNX format and was run
using the ONNX Inference session. Moreover, the onnx model was then
evaluated by measuring its latency and average precision and these
benchmarks were compared to those of the original model. Moreover,
changing the model weights, inputs and targets to the 16 bit float
data type instead of 32 bit floats was evaluated as well.